run_make_support: move fs helpers to own module
